flyttet formatTime til datFnsUtils og renamet den til formatMilliseconds
